+++ src/main/java/org/auth/csd/datalab/services/PlacementService.java
@@ -18,6 +18,7 @@ import org.auth.csd.datalab.common.models.values.MetaMetric;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.*;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
@@ -70,6 +71,55 @@ public class PlacementService implements PlacementInterface {
         events.localListen(localListener, EventType.EVT_NODE_JOINED, EventType.EVT_NODE_FAILED);
     }
 
+    private void clearPreviousUnstable(String hostname){
+        //Set replication on the node to false, so it stops replicating data
+        MovementInterface srvInterface = ignite.services(ignite.cluster().forHost(hostname)).serviceProxy(MovementInterface.SERVICE_NAME, MovementInterface.class, false);
+        srvInterface.setReplication(false);
+        //Remove the entries from the replicas cache
+        HashSet<String> sourceSet = new HashSet<>();
+        sourceSet.add(hostname);
+        HashMap<HostMetricKey, MetaMetric> tmpMmetrics = srvInterface.extractMonitoringList(new HashMap<>(), sourceSet);
+        if(!tmpMmetrics.isEmpty()) myReplica.removeAll(tmpMmetrics.keySet());
+        //Update meta
+        Restarts nodeInfo = myRestarts.get(hostname);
+        //And update the replica set
+        nodeInfo.replicaTo.forEach(k -> {
+            Restarts tmpDest = myRestarts.get(k);
+            tmpDest.removeSource(hostname);
+            myRestarts.put(k, tmpDest);
+        });
+        nodeInfo.clearDestination();
+        myRestarts.put(hostname, nodeInfo);
+    }
+
+    private void clearPreviousStable(String hostname){
+        Restarts nodeInfo = myRestarts.get(hostname);
+        nodeInfo.replicaFrom.forEach(k -> {
+            Restarts tmpRestart = myRestarts.get(k);
+            MovementInterface srvInterface = ignite.services(ignite.cluster().forHost(k)).serviceProxy(MovementInterface.SERVICE_NAME, MovementInterface.class, false);
+            HashSet<String> sourceSet = new HashSet<>();
+            sourceSet.add(k);
+            HashMap<HostMetricKey, MetaMetric> tmpMmetrics = srvInterface.extractMonitoringList(new HashMap<>(), sourceSet);
+            AtomicBoolean continueReplication = new AtomicBoolean(false);
+            if(!tmpMmetrics.isEmpty()) {
+                tmpMmetrics.keySet().forEach(metric -> {
+                    Set<String> tmpDests = myReplica.get(metric);
+                    tmpDests.remove(hostname);
+                    if(tmpDests.isEmpty()) myReplica.remove(metric);
+                    else {
+                        myReplica.put(metric, tmpDests);
+                        continueReplication.set(true);
+                    }
+                });
+            }
+            srvInterface.setReplication(continueReplication.get());
+            tmpRestart.removeDestination(hostname);
+            myRestarts.put(k, tmpRestart);
+        });
+        nodeInfo.clearSource();
+        myRestarts.put(hostname, nodeInfo);
+    }
+
     private boolean decidePlacement() {
         //First find the top-N problematic nodes
         List<NodeScore> problematicNodes = findProblematicNodes(1);
@@ -87,7 +137,11 @@ public class PlacementService implements PlacementInterface {
                         //Update the last replication time
                         Restarts tmp = myRestarts.get(source.node); //No need for check. It has been completed on finding the problematic node
                         tmp.addReplication(currTime);
+                        tmp.addDestination(dest.getKey().node);
                         myRestarts.put(source.node, tmp);
+                        Restarts tmp2 = myRestarts.get(dest.getKey().node); //No need for check. It has been completed on finding the stable node
+                        tmp2.addSource(source.node);
+                        myRestarts.put(dest.getKey().node, tmp2);
                         //Update the replica table
                         String newHost = dest.getKey().node;
                         dest.getValue().forEach(k -> {
@@ -97,9 +151,9 @@ public class PlacementService implements PlacementInterface {
                             repls.add(newHost);
                             myReplica.put(k, repls);
                         });
-                        return true;
                     }
                 }
+                return true;
             }
         }
         return false;
@@ -111,7 +165,13 @@ public class PlacementService implements PlacementInterface {
         for (ClusterNode node : ignite.cluster().forServers().forRemotes().nodes()) {
             String hostname = node.hostNames().iterator().next();
             Restarts nodeFails = (myRestarts.containsKey(hostname)) ? myRestarts.get(hostname) : null;
-            if (nodeFails == null || nodeFails.failTimes.size() < 1 || nodeFails.lastReplication >= currTime - PROBLEMATIC_TAG_INTERVAL) continue; //It is not a problematic node
+            if(nodeFails == null) {
+                myRestarts.put(hostname, new Restarts(currTime));
+                continue;
+            }
+            if (nodeFails.failTimes.isEmpty() || nodeFails.lastReplication >= (currTime - (PROBLEMATIC_TAG_INTERVAL*1000))) {
+                continue; //It is not a problematic node
+            }
             //Check if it reached max capacity
             if(getReplicatedMetricsMax(hostname) >= MAX_OUT_REPLICAS) continue;
             //Else get a weighted average of the time between fails
@@ -127,11 +187,15 @@ public class PlacementService implements PlacementInterface {
                 if (timestamp <= currTime - OLD_PERIOD) a = 0.5 * oldFails;
                 else if (timestamp <= currTime - MID_PERIOD) a = midFails;
                 else a = 2D * newFails;
-                if (i == 0) sum += a / (timestamp - startTime);
-                else sum += a / (timestamp - nodeFails.failTimes.get(i - 1));
+                if (i == 0) sum += a / ((timestamp - startTime)/1000D);
+                else sum += a / ((timestamp - nodeFails.failTimes.get(i - 1))/1000D);
             }
             //Add the node with the score in the map if it passes above a threshold
-            if (sum >= PROBLEMATIC_THRESHOLD) problematicNodes.add(new NodeScore(hostname, sum));
+            if (sum >= PROBLEMATIC_THRESHOLD) {
+                problematicNodes.add(new NodeScore(hostname, sum));
+                //check if a problematic node is a replica destination for another node
+                if(!nodeFails.replicaFrom.isEmpty()) clearPreviousStable(hostname);
+            }
         }
         //Find and return the top N nodes
         if (problematicNodes.isEmpty()) return problematicNodes;
@@ -144,15 +208,19 @@ public class PlacementService implements PlacementInterface {
         for (ClusterNode node : ignite.cluster().forServers().forRemotes().nodes()) {
             String hostname = node.hostNames().iterator().next();
             Restarts nodeInfo = (myRestarts.containsKey(hostname)) ? myRestarts.get(hostname) : null;
-            if(!exclude.contains(hostname) && nodeInfo != null){
-                //Get number of fails
+            if(nodeInfo == null) continue;
+            if(!exclude.contains(hostname)){
+                    //Get number of fails
                 long nodeFails = nodeInfo.failTimes.stream().filter(k -> k > currTime - MID_PERIOD).count();
-                if (nodeFails > STABLE_MAX_FAILS || nodeInfo.lastReplication > currTime - PROBLEMATIC_TAG_INTERVAL) //Either it has failed too many times or it has been replicated in the last hour
+                if (nodeFails > STABLE_MAX_FAILS || nodeInfo.lastReplication > currTime - (PROBLEMATIC_TAG_INTERVAL*1000)) {//Either it has failed too many times or it has been replicated in the last hour
                     continue; //If it failed many times recently exclude it
+                }
                 //Sum up the restarts and the fails
-                double nodeRestarts = (-nodeInfo.restartTimes.size() - (nodeFails * 3D)) * (currTime - nodeInfo.startTime); //Get the time alive of the node minus the restarts
+                double nodeRestarts = (-nodeInfo.restartTimes.size() - (nodeFails * 3D)) * ((currTime - nodeInfo.startTime)/1000D); //Get the time alive of the node minus the restarts
                 //Add the node with the score in the map (Bigger score = better node)
                 stableNodes.add(new NodeScore(hostname, nodeRestarts));
+                //If the node was unstable in the past remove the replications
+                if(!nodeInfo.replicaTo.isEmpty()) clearPreviousUnstable(hostname);
             }
         }
         //Find and return the top N nodes